+++ src/config.py
@@ -11,8 +11,8 @@ VEKN_VTES_ENCODING = "utf-8"
 #: where to store our data
 TWDA_FILE = path.join(path.dirname(__file__), "TWDA.pkl")
 VTES_FILE = path.join(path.dirname(__file__), "VTES.pkl")
-CARDS_RULINGS_FILE = path.join(path.dirname(__file__), "cards_rulings.yaml")
-GENERAL_RULINGS_FILE = path.join(path.dirname(__file__), "general_rulings.yaml")
+CARDS_RULINGS_FILE = path.join(path.dirname(__file__), "cards-rulings.yaml")
+GENERAL_RULINGS_FILE = path.join(path.dirname(__file__), "general-rulings.yaml")
 RULINGS_LINK_FILE = path.join(path.dirname(__file__), "rulings-links.yaml")
 #: classic headers in deck lists
 HEADERS = [